+++ pkg/service/todo_item.go
@@ -18,7 +18,7 @@ func NewToDoItemService(repo repository.ToDoItem, listRepo repository.ToDoList) 
 	}
 }
 
-func (s *ToDoItemService) Create(userId, listId int, item todo.ToDoItem) (int, error) {
+func (s *ToDoItemService) Create(userId, listId int64, item todo.ToDoItem) (int64, error) {
 	_, err := s.ListRepo.GetById(userId, listId)
 	if err != nil {
 		return 0, errors.New("there is no list with such id, or user doesn't have permission for it")
@@ -26,7 +26,7 @@ func (s *ToDoItemService) Create(userId, listId int, item todo.ToDoItem) (int, e
 	return s.Repo.Create(listId, item)
 }
 
-func (s *ToDoItemService) GetAll(userId, listId int) ([]todo.ToDoItem, error) {
+func (s *ToDoItemService) GetAll(userId, listId int64) ([]todo.ToDoItem, error) {
 	_, err := s.ListRepo.GetById(userId, listId)
 	if err != nil {
 		return nil, errors.New("there is no list with such id, or user doesn't have permission for it")
@@ -36,15 +36,15 @@ func (s *ToDoItemService) GetAll(userId, listId int) ([]todo.ToDoItem, error) {
 
 }
 
-func (s *ToDoItemService) GetById(userId, itemId int) (todo.ToDoItem, error) {
+func (s *ToDoItemService) GetById(userId, itemId int64) (todo.ToDoItem, error) {
 	return s.Repo.GetById(userId, itemId)
 }
 
-func (s *ToDoItemService) Delete(userId, itemId int) error {
+func (s *ToDoItemService) Delete(userId, itemId int64) error {
 	return s.Repo.Delete(userId, itemId)
 }
 
-func (s *ToDoItemService) Update(userId, listId int, updateData todo.UpdateItemInput) error {
+func (s *ToDoItemService) Update(userId, listId int64, updateData todo.UpdateItemInput) error {
 	if err := updateData.Validate(); err != nil {
 		return err
 	}
