+++ pkg/service/todo_list.go
@@ -15,23 +15,23 @@ func NewToDoListService(repo repository.ToDoList) *ToDoListService {
 	}
 }
 
-func (s *ToDoListService) Create(userId int, list todo.ToDoList) (int, error) {
+func (s *ToDoListService) Create(userId int64, list todo.ToDoList) (int64, error) {
 	return s.repo.Create(userId, list)
 }
 
-func (s *ToDoListService) GetAll(userId int) ([]todo.ToDoList, error) {
+func (s *ToDoListService) GetAll(userId int64) ([]todo.ToDoList, error) {
 	return s.repo.GetAll(userId)
 }
 
-func (s *ToDoListService) GetById(userId, listId int) (todo.ToDoList, error) {
+func (s *ToDoListService) GetById(userId, listId int64) (todo.ToDoList, error) {
 	return s.repo.GetById(userId, listId)
 }
 
-func (s *ToDoListService) Delete(userId, listId int) error {
+func (s *ToDoListService) Delete(userId, listId int64) error {
 	return s.repo.Delete(userId, listId)
 }
 
-func (s *ToDoListService) Update(userId, listId int, updateData todo.UpdateListInput) error {
+func (s *ToDoListService) Update(userId, listId int64, updateData todo.UpdateListInput) error {
 	if err := updateData.Validate(); err != nil {
 		return err
 	}
